Implemented xmlcatalog stuff: xmlcatalog is required (no big problem, because we already need libxml2 for building bluefish). We need to it write our own catalog.xml, which we install into $datadir/xml/$PACKAGE besides the DTD. And it's necessary to update the systems XML catalog, which is optional. Therefor the user gets two now configure options:
 --enable-xml-catalog-update: The user can choose to disable the catalog update.
 --with-xml-catalog: The user should be able to determine the systems XML catalog.

Depending, if the update is enabled and if the given (or standard /etc/xml/catalog) file is avaialble, the update is run during installation.

Added the installation of the DTD and the catalog to the Makefile. Fixed a small bug: better install pascal pattern and do not install the sample.